refactor: Move internal log package to PAM module
It's only used there now.
