+++ pam/internal/adapter/gdmmodel.go
@@ -10,8 +10,8 @@ import (
 	"github.com/msteinert/pam/v2"
 	"github.com/ubuntu/authd"
 	"github.com/ubuntu/authd/internal/brokers"
-	"github.com/ubuntu/authd/internal/log"
 	"github.com/ubuntu/authd/pam/internal/gdm"
+	"github.com/ubuntu/authd/pam/internal/log"
 	"github.com/ubuntu/authd/pam/internal/proto"
 )
