+++ pam/internal/dbusmodule/transaction_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/godbus/dbus/v5"
 	"github.com/msteinert/pam/v2"
 	"github.com/stretchr/testify/require"
-	"github.com/ubuntu/authd/internal/log"
 	"github.com/ubuntu/authd/internal/testutils"
 	"github.com/ubuntu/authd/pam/internal/dbusmodule"
+	"github.com/ubuntu/authd/pam/internal/log"
 )
 
 const ifaceName = "com.ubuntu.authd.pam"
